requirements.txt
# Telegram
pyTelegramBotAPI>=4.14.0

# Google GenAI SDK (الأحدث - ليس google-generativeai)
google-genai>=1.0.0

# Database: SQLite3 (built-in) + FTS5

# File processing
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
